JDK level defined in the pom.xml file, API package moved
